Add dressed state transformation for U_j and finalize simulation
This commit adds functionality to work with dressed states for the
computation of the U_j matrices (previously with nans) in the quantum
dot system, including computing dressed energies and states,
transforming to a rotating frame using dressed states, and adding a
timer class for performance measurement. It also adds a separate
function to only plot the mean population trajectories (no individual
trajectories).

Next:
- define control window (with finer dt) and simulation window
- do optimal control
- implement actuator dynamics (notch filter)